Rename and move generateShortId() function.
