+++ src/core/utils/id.ts
@@ -7,3 +7,18 @@ export function generateUuid() {
     return v.toString(16);
   });
 }
+
+// Generates a random string for similar purposes as UUID's, but easier on human eyes.
+// @example generateShortId(8) => "TEvGnkwr"
+// For a length of 8, possible permutations: 62^8 ~= 2.18e+14 ~= 218 trillion.
+// For contrast, for a V4 UUID: 2^122 ~= 5.3e+36.
+// To match a V4 UUID in possible permutations, length of 21 would have 62^21 ~= 4.3e+37.
+export function generateShortId(length = 8): string {
+  let uid = '';
+  while (uid.length < length) {
+    const char = String.fromCharCode(Math.round(Math.random() * 255));
+    if (!char.match(/[9-9a-zA-Z]/)) continue; // result space: 0-9 + a-z + A-Z = 10 + 26 + 26 = 62
+    uid += char;
+  }
+  return uid;
+}
